requirements/base.txt
python-slugify==8.0.4  # https://github.com/un33k/python-slugify
Pillow==11.1.0  # https://github.com/python-pillow/Pillow
rcssmin==1.1.2 # https://github.com/ndparker/rcssmin
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
redis==5.2.1  # https://github.com/redis/redis-py
hiredis==3.1.0  # https://github.com/redis/hiredis-py
celery==5.4.0  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.7.0  # https://github.com/celery/django-celery-beat
flower==2.0.1  # https://github.com/mher/flower
# Async configuration
uvicorn[standard]==0.34.0  # https://github.com/encode/uvicorn
uvicorn-worker==0.3.0  # https://github.com/Kludex/uvicorn-worker
deepl==1.22.0
django-modeltrans==0.8.0

# Django
# ------------------------------------------------------------------------------
django>=3.2,<4.3 # pyup: < 5.1  # https://www.djangoproject.com/
django-environ==0.12.0  # https://github.com/joke2k/django-environ
django-model-utils==5.0.0  # https://github.com/jazzband/django-model-utils
django-allauth[mfa]==65.8.1  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.3  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==2024.10  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-compressor==4.5.1  # https://github.com/django-compressor/django-compressor
django-libsass==0.9
django-redis==5.4.0  # https://github.com/jazzband/django-redis
django-tinymce==4.1.0
django-filebrowser==4.0.3
django-grappelli==3.0.10
django-flatpickr==2.0.3
# Django Oscar
# ------------------------------------------------------------------------------
django-oscar==3.2.5
setuptools==78.1.1
phonenumbers==9.0.2
django-treebeard==4.7.1
django-widget-tweaks==1.5.0
django-extra-views==0.14.0
django-haystack==3.3.0
django-phonenumber-field==6.4.0
purl==1.6
pycountry==24.6.1
pysolr==3.10.0
sorl-thumbnail>=12.10.0,<13.0.0
